+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   - Feat: add `ITrimmedLexer/IReadonlyTrimmedLexer`.
   - Feat: `Builder.append` can accept actions with different data bindings in one call.
   - Feat: add `Lexer.javascript.evalStringContent/simpleStringLiteral`.
-  - Feat: add `Lexer.binaryIntegerLiteral`.
+  - Feat: add `Lexer.integerLiteral/binaryIntegerLiteral/octalIntegerLiteral/hexIntegerLiteral`.
   - Feat: add `Lexer.json`.
   - Feat: add `Lexer.SubAction`, `Lexer.str2subAction/regex2subAction`.
   - Fix: `Lexer.trimStart` shouldn't collect non-muted errors.